+++ app/src/main/java/amon/pramhathai/sasiporn/rmutsv/ac/th/rubbershop/fragment/CustomerReportSheetFragment.java
@@ -99,7 +99,7 @@ public class CustomerReportSheetFragment extends Fragment {
             totalADouble = totalADouble + Double.parseDouble(balanceStrings[i]);
         }
 
-        result = Double.toString(totalADouble);
+        result = Integer.toString((int) totalADouble);
 
         return result;
     }
