+++ app/src/main/java/amon/pramhathai/sasiporn/rmutsv/ac/th/rubbershop/fragment/DepositFragment.java
@@ -80,9 +80,6 @@ public class DepositFragment extends Fragment {
             String[] balanceStrings = new String[jsonArray.length()];
             String[] nameCustomerStrings = new String[jsonArray.length()];
 
-
-
-
             for (int i = 0; i<jsonArray.length(); i+=1) {
                 JSONObject jsonObject = jsonArray.getJSONObject(i);
 
@@ -143,7 +140,7 @@ public class DepositFragment extends Fragment {
             totalADouble = totalADouble + Double.parseDouble(balanceStrings[i]);
         }
 
-        result = Double.toString(totalADouble);
+        result = Integer.toString((int) totalADouble);
 
         return result;
     }
