+++ app/src/main/java/amon/pramhathai/sasiporn/rmutsv/ac/th/rubbershop/fragment/LatexRubberFragment.java
@@ -142,51 +142,6 @@ public class LatexRubberFragment extends Fragment {
 
     }
 
-//    public void shareScreenShotM(View view, NestedScrollView scrollView) {
-//
-//        Bitmap bm = takeScreenShort(view, scrollView);
-//        File file = savePic(bm);
-//    }
-//
-//    public Bitmap takeScreenShort(View u, NestedScrollView z) {
-//        u.setDrawingCacheEnabled(true);
-//        int totalHeight = z.getChildAt(0).getHeight();
-//        int totalWidth = z.getChildAt(0).getWidth();
-//
-//        Log.d("yoheight", " " + totalHeight);
-//        Log.d("yowidth", " " + totalWidth);
-//        u.layout(0, 0, totalWidth, totalHeight);
-//        u.buildDrawingCache();
-//        Bitmap b = Bitmap.createBitmap(u.getDrawingCache());
-//        u.setDrawingCacheEnabled(false);
-//        u.destroyDrawingCache();
-//        return b;
-//    }
-//
-//    public static File savePic(Bitmap bm) {
-//        ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-//        bm.compress(Bitmap.CompressFormat.JPEG, 100, bytes);
-//        File sdCardDirectory = new File(Environment.getExternalStorageDirectory() + "/Foldername");
-//
-//        if (!sdCardDirectory.exists()) {
-//            sdCardDirectory.mkdirs();
-//        }
-//
-//        File file = null;
-//        try {
-//            file = new File(sdCardDirectory, Calendar.getInstance()
-//                    .getTimeInMillis() + ".jpg");
-//            file.createNewFile();
-//            new FileOutputStream(file).write(bytes.toByteArray());
-//            Log.d("Fabsolute", "File Saved:: ==>" + file.getAbsolutePath());
-//            Log.d("Sabsolute", "File Saved:: ==>" + sdCardDirectory.getAbsolutePath());
-//        } catch (IOException e) {
-//            e.printStackTrace();
-//        }
-//        return file;
-//
-//    }
-
     private void addBuyRuber() {
         try {
             statusABoolean = false;
@@ -218,7 +173,7 @@ public class LatexRubberFragment extends Fragment {
 
                     double dryRubber = Double.parseDouble(dryRubberString);
                     double totalPrice = dryRubber * Double.parseDouble(priceString);
-                    totalString = Double.toString(totalPrice);
+                    totalString = Integer.toString((int) totalPrice);
 
                     TextView textView = getView().findViewById(R.id.txtTotal);
                     textView.setText(totalString);
